my-k8s-data-app/reader/requirements.txt
Flask
mysql-connector-python
prometheus_client
